INJIWEB-1851:Move latest ui automation changes from develop to release-0.16.x
