+++ ui-test/src/main/java/io/inji/testrig/apirig/injiweb/testscripts/AddIdentity.java
@@ -1,4 +1,4 @@
-package io.mosip.testrig.apirig.injiweb.testscripts;
+package io.inji.testrig.apirig.injiweb.testscripts;
 
 import java.lang.reflect.Field;
 import java.text.DateFormat;
@@ -38,6 +38,7 @@ import io.mosip.testrig.apirig.utils.KernelAuthentication;
 import io.mosip.testrig.apirig.utils.OutputValidationUtil;
 import io.mosip.testrig.apirig.utils.ReportUtil;
 import io.mosip.testrig.apirig.utils.RestClient;
+import io.mosip.testrig.apirig.utils.SecurityXSSException;
 import io.restassured.response.Response;
 
 public class AddIdentity extends AdminTestUtil implements ITest {
@@ -78,7 +79,7 @@ public class AddIdentity extends AdminTestUtil implements ITest {
 	 * @throws AdminTestException
 	 */
 	@Test(dataProvider = "testcaselist")
-	public void test(TestCaseDTO testCaseDTO) throws AuthenticationTestException, AdminTestException {
+	public void test(TestCaseDTO testCaseDTO) throws AuthenticationTestException, AdminTestException, SecurityXSSException {
 		testCaseName = testCaseDTO.getTestCaseName();
 		if (HealthChecker.signalTerminateExecution) {
 			throw new SkipException(
