+++ ui-test/src/main/java/runnerfiles/Runner.java
@@ -27,7 +27,7 @@ import java.util.List;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(
-		features = {"/home/mosip/featurefiles/"},
+		features = {"/home/inji/featurefiles/"},
 		dryRun = false,
 		glue = {"stepdefinitions", "utils"},
 		snippets = SnippetType.CAMELCASE,
@@ -116,7 +116,7 @@ public class Runner extends AbstractTestNGCucumberTests{
 			LOGGER.info("IDE :" + homeDir);
 		} else {
 			File dir = new File(System.getProperty("user.dir"));
-			homeDir = new File(dir.getParent() + "/mosip/testNgXmlFiles");
+			homeDir = new File(dir.getParent() + "/inji/testNgXmlFiles");
 			LOGGER.info("ELSE :" + homeDir);
 		}
 		File[] files = homeDir.listFiles();
@@ -220,7 +220,7 @@ public class Runner extends AbstractTestNGCucumberTests{
 	    if (os.contains("windows")) {
 	        System.setProperty("cucumber.features", "src\\test\\resources\\featurefiles\\");
 	    } else {
-	        System.setProperty("cucumber.features", "/home/mosip/featurefiles/");
+	        System.setProperty("cucumber.features", "/home/inji/featurefiles/");
 	    }
 	}
