+++ ui-test/src/main/resources/config/injiweb.properties
@@ -1,7 +1,83 @@
-#------------------------- DB Connectivity proprties ------------------------#
-#-- When test rig ran in docker,few of below dervied from the environment ---#
-apiEnvUser=api-internal.released
+#---------------------------------- End point(s) relative URLs ----------------------------------#
 apiInternalEndPoint=https://api-internal.released.mosip.net
+actuatorMimotoEndpoint=/v1/mimoto/actuator/env
 injiWebUi=https://injiweb.qa-inji1.mosip.net/
 InsuranceUrl=https://registry.released.mosip.net/api/v1/Insurance
-actuatorMimotoEndpoint=/v1/mimoto/actuator/env
+
+
+#---------------------------------- Modifiable Properties ----------------------------------------#
+
+#------------------------ Environment URLs and Database Connections ------------------------------#
+
+# Keycloak URL.
+keycloak-external-url=https://iam.released.mosip.net
+
+# PostgreSQL URLs for audit and partner databases.
+audit_url=jdbc:postgresql://released.mosip.net:5432/mosip_audit
+partner_url=jdbc:postgresql://released.mosip.net:5432/mosip_ida
+
+# Database server for connections.
+db-server=released.mosip.net
+
+
+#--------------------------------- secrets and passwords  -----------------------------------------#
+
+#---------------------------------  Keycloak Passwords --------------------------------------------#
+# Used for Keycloak authentication.
+keycloak_Password=
+
+#-------------------------------- PostgreSQL Database Passwords -----------------------------------#
+# Credentials for connecting to Postgres databases.
+audit_password=
+partner_password=
+postgres-password=
+
+#-------------------------------- Client Secret Keys ----------------------------------------------#
+# These keys are used for various services, make sure to update the values as required when running locally.
+
+mosip_partner_client_secret=
+mosip_pms_client_secret=
+mosip_resident_client_secret=
+mosip_idrepo_client_secret=
+mosip_reg_client_secret=
+mosip_admin_client_secret=
+mosip_hotlist_client_secret=
+mosip_regproc_client_secret=
+mpartner_default_mobile_secret=
+mosip_testrig_client_secret=
+AuthClientSecret=
+
+
+#---------------------------- Generic Configuration ----------------------------------------------#
+
+# Enable or disable debugging mode (yes/no).
+
+enableDebug=no
+
+# Mock Notification Channels (email/phone/email,phone).
+mockNotificationChannel=email,phone
+
+#------------------------ Mosip Components Base URLs ---------------------------------------------#
+# Define base URLs for different components if required.
+# Example: 
+# mosip_components_base_urls = auditmanager=api-internal.released.mosip.net;idrepository=api-internal.released.mosip.net;authmanager=api-internal.released.mosip.net;resident=api-internal.released.mosip.net;partnermanager=api-internal.released.mosip.net;idauthentication=api-internal.released.mosip.net;masterdata=api-internal.released.mosip.net;idgenerator=api-internal.released.mosip.net;policymanager=api-internal.released.mosip.net;preregistration=api-internal.released.mosip.net;keymanager=api-internal.released.mosip.net;mock-identity-system=api.released.mosip.net
+# Feel free to add more components as needed.
+mosip_components_base_urls =
+
+
+#-------------------------------- Module Name Pattern --------------------------------------------#
+# Define module name pattern if required.
+# Example: 
+# moduleNamePattern = (mimoto|resident)
+# Feel free to add more values as needed.
+moduleNamePattern =
+
+
+#-------------------------------- Uncomment for Local Run ----------------------------------------#
+
+# Path to the authentication certificates (if running locally, uncomment the below line and keep the value empty).
+# authCertsPath =
+
+# X-XSS-Protection: Controls the XSS (Cross-Site Scripting) filter in browsers.
+# Values: (yes/no)
+xssProtectionCheck = no
