+++ ui-test/src/test/resources/config.properties
@@ -1,5 +1,5 @@
-MOSIP_INJIWEB_GOOGLE_CLIENT_ID=
-MOSIP_INJIWEB_GOOGLE_CLIENT_SECRET=
-MOSIP_INJIWEB_GOOGLE_REFRESH_TOKEN=
+INJIWEB_GOOGLE_CLIENT_ID=
+INJIWEB_GOOGLE_CLIENT_SECRET=
+INJIWEB_GOOGLE_REFRESH_TOKEN=
 issuerSearchText=National Identity Department
 issuerSearchTextforSunbird=StayProtected Insurance
